requirements.txt
gradio
pandas
plotly
python-dotenv